utils/slack: split files
